fix(storage): fix local storage bucket about

+++ pkg/storage/local/impl.go
@@ -5,6 +5,7 @@ import (
 	"io"
 	"os"
 	"path/filepath"
+	"strconv"
 	"strings"
 	"sync/atomic"
 
@@ -17,12 +18,18 @@ const (
 
 func (s *Storage) Put() (storage.Writer, error) {
 	id := s.genFileID()
-	key := makeFileKey(s.prefix, id)
-	return NewWriter(s.base, key)
+	key := packFileKey(s.prefix, id)
+	bucket := genBucket(id)
+	return NewWriter(s.base, bucket, key)
 }
 
 func (s *Storage) Get(key string) (r io.ReadCloser, e error) {
-	fp := filepath.Join(s.base, key)
+	id, e := parseFileKey(key)
+	if e != nil {
+		return nil, e
+	}
+	bucket := genBucket(id)
+	fp := filepath.Join(s.base, bucket, key)
 	f, e := os.Open(fp)
 	if e != nil {
 		return nil, e
@@ -31,14 +38,36 @@ func (s *Storage) Get(key string) (r io.ReadCloser, e error) {
 }
 
 func (s *Storage) Delete(key string) (e error) {
-	fp := filepath.Join(s.base, key)
+	id, e := parseFileKey(key)
+	if e != nil {
+		return e
+	}
+	bucket := genBucket(id)
+	fp := filepath.Join(s.base, bucket, key)
 	return os.Remove(fp)
 }
 
-func makeFileKey(base string, id uint32) string {
+func packFileKey(base string, id uint32) string {
 	idStr := fmt.Sprintf("%08X", id)
 	return strings.Join([]string{idStr, base}, fileKeySplit)
 }
+
+func parseFileKey(key string) (uint32, error) {
+	vec := strings.SplitN(key, fileKeySplit, 2)
+	if len(vec) != 2 {
+		return 0, fmt.Errorf("bad file key: \"%v\"", key)
+	}
+	id, e := strconv.Atoi(vec[0])
+	if e != nil {
+		return 0, fmt.Errorf("bad file key: %v", e)
+	}
+	return uint32(id), nil
+}
+
 func (s *Storage) genFileID() uint32 {
 	return atomic.AddUint32(&s.lastKey, 1)
+}
+
+func genBucket(id uint32) string {
+	return fmt.Sprintf("%04X", id%bucketNum)
 }
